+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/navigation/SetBirthDateNavigation.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.auth.presentation.navigation
+package com.teamoffroad.feature.auth.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
