+++ feature/main/src/main/java/com/teamoffroad/feature/main/MainNavigator.kt
@@ -10,10 +10,10 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
 import com.teamoffroad.core.navigation.Route
-import com.teamoffroad.feature.auth.presentation.navigation.navigateToSetBirthDate
-import com.teamoffroad.feature.auth.presentation.navigation.navigateToSetCharacter
-import com.teamoffroad.feature.auth.presentation.navigation.navigateToSetGender
-import com.teamoffroad.feature.auth.presentation.navigation.navigateToSetNickname
+import com.teamoffroad.feature.auth.navigation.navigateToSetBirthDate
+import com.teamoffroad.feature.auth.navigation.navigateToSetCharacter
+import com.teamoffroad.feature.auth.navigation.navigateToSetGender
+import com.teamoffroad.feature.auth.navigation.navigateToSetNickname
 import com.teamoffroad.feature.explore.presentation.navigation.navigateExplore
 import com.teamoffroad.feature.home.navigation.navigateHome
 import com.teamoffroad.feature.mypage.navigation.navigateMypage
